Decompose dependencies around utils/data.js to avoid circular dependencies.
