+++ client/src/components/FilesDialog/FilesInput.vue
@@ -3,7 +3,7 @@ import { BFormInput } from "bootstrap-vue";
 import { computed } from "vue";
 
 import { type FileSourceBrowsingMode, type FilterFileSourcesOptions } from "@/api/remoteFiles";
-import { filesDialog } from "@/utils/data";
+import { filesDialog } from "@/utils/dataModals";
 
 import { type SelectionItem } from "../SelectionDialog/selectionTypes";
