+++ client/src/components/Upload/RulesInput.vue
@@ -5,7 +5,7 @@ import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { getGalaxyInstance } from "app";
 import { BAlert, BButton } from "bootstrap-vue";
 import { getRemoteEntries, getRemoteEntriesAt } from "components/Upload/utils";
-import { filesDialog } from "utils/data";
+import { filesDialog } from "utils/dataModals";
 import { urlData } from "utils/url";
 import { computed, ref } from "vue";
